% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Traffic accidents in 2013.}
\format{
A data frame with 30202 rows and 50 variables.
}
\source{
\url{https://d3c33hcgiwev3.cloudfront.net/_e1adac2a5f05192dc8780f3944feec13_fars_data.zip?Expires=1639440000&Signature=C64jowmPeR0f1jkssP2rqXViiLUi2L4hL3aEYLrxarR2es0~oIkSC6TL90wA6QuP7K8QPDpRBWKDpANQlJBWIWzUCxpz9KzZ4-i-EUjTVEd6o37~1MzmFmn0NWrOAUuEMVryocTNvZugD0AlJDOMVJ6hTwlZvk6w92NCR-qyJGw_&Key-Pair-Id=APKAJLTNE6QMUY6HBC5A}
}
\usage{
accident_2013
}
\description{
Data from the US National Highway Traffic Safety Administration's Fatality
Analysis Reporting System, which is a nationwide census providing the
American public yearly data regarding fatal injuries suffered in motor
vehicle traffic crashes.
}
\keyword{datasets}
