% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The code (number) of the state for which to generate a plot.}

\item{year}{Could be a single string or a list of years.}
}
\value{
This functions returns a map with points that represent car accidents
in the selected state during the given year.
}
\description{
`fars_map_state` function generates a plot of accidents recorded across the
selected state during the given year.
}
\examples{
\dontrun{fars_map_state(state.num = 1, year = 2015)}

}
