% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read input file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that represents the name of the input file.}
}
\value{
The function returns a data frame of the provided data. In case that
        there is no file with the provided name, the following error message
        will be shown: "file XYZ does not exist".
}
\description{
`fars_read` function reads a single archive file based on the provided filename.
Filename should be generated with the make_filename function.
The archive file (CSV) contains the data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a nationwide
census providing the American public yearly data regarding fatal injuries
suffered in motor vehicle traffic crashes.
}
\examples{
\dontrun{fars_read(filename = "accident_2015.csv.bz2")}

}
