% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Could be a single string or a list of years.}
}
\value{
The function returns a list of elements, where each element
        contains two columns (MONTH and year). If you pass an incorrect
        year (a year for which no data is available) to the function, it will
        return the corresponding warning message.
}
\description{
`fars_read_years` function utilizes the make_filename() as well as the
fars_read() functions to import years. Under the hood the function performs
some data manipulation (mutation, selection) and returns the list of the
resulting tibbles.
}
\examples{
\dontrun{fars_read_years(years = 2015)}
\dontrun{fars_read_years(years = c(2014, 2015))}

}
