% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The only input of the function. It should be a positive number.}
}
\value{
This function returns a string representing the name of the file.
        The resulting filename will be in the following format:
        accident_YEAR.csv.bz2 .
}
\description{
`make_filename` function generates the filename string that will be used in
the fars_read function.
}
\examples{
\dontrun{make_filename(year = 2015)}

}
